refactor(cyfront.c): reformat code inside cyfront.c
Reformat code inside cyfront.c, because it is easier to refactor the
whole thing if it is properly formatted. That is the first step to
achieve clean code.

This file uses pretty uggly spaghetti code, and it needs to be
refactored into proper modern c code. On the other hand, it can be used
to reverse engineer the original gentle source code of the gentle
compiler kit.